change leader class into thread safe
